pyproject.toml
[project]
name = "nowagift"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama==0.4.6",
    "decorator==5.2.1",
    "google-genai>=1.32.0",
    "imageio==2.37.0",
    "imageio-ffmpeg==0.6.0",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.6",
    "moviepy==2.2.1",
    "numpy==2.3.2",
    "pillow==11.3.0",
    "proglog==0.1.12",
    "python-dotenv==1.1.1",
    "streamlit>=1.48.1",
    "tqdm==4.67.1",
]
